requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
apify==3.0.5
apify-client==2.3.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
google-api-python-client==2.150.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.23.1
pdfminer.six
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
